+++ packages/publisher/src/web/views/server.ts
@@ -1,0 +1,12 @@
+export { ErrorView } from "./ErrorView.js";
+export type { ErrorViewProps } from "./ErrorView.js";
+
+export { AuthView } from "./AuthView.js";
+export type { AuthViewProps } from "./AuthView.js";
+
+export { MarketingView } from "./MarketingView.server.js";
+export type {
+  MarketingFaqItem,
+  MarketingFeatureItem,
+  MarketingViewProps,
+} from "./MarketingView.js";